Created Ewon library release branch action
This github action will create a release branch for all versioned
tags. If a tag is pushed that starts with v a new branch will be Created
with the branch name release/X.Y where X is the major version number
and y is the minor version number (taken from the tag name).